Move version to be config-wide, not specific tests
This change moves the version to be configuration-wide, so it applies to
the entire run and test of results instead of putting it on every
individual test. This introduces a separate configuration area called
`details` which is where name and version now live, with `defaults`
acting as per-test values. This also rearranged a few things to make
this work easier.

This needs tests, badly.