Organize code into better module/script split
As this is an executable script, it should really only expose the
smallest bit possible to execute itself. This change hides the guts of
the code in the `_pine` module and now installs only a small piece of
code as the bin/pine executable which just imports _pine and runs main
from there.